Удалил пустые структуры из ответов Разбил модели на отдельные файлы
