+++ internal/client/cap/models/admin.go
@@ -1,0 +1,11 @@
+package models
+
+type AdminCheckRequestBody struct {
+	UserName string `json:"username"`
+	Password string `json:"password"`
+}
+
+type AdminCheckResponseBody struct {
+	Token        string `json:"token"`
+	RefreshToken string `json:"refreshToken"`
+}
